Adjust the handling of gv input to allow parsing gprof2dot output
gprof2dot is a program that can take output from gprof, callgrind, and
perf and produce a dot output file.  The format is different from what
sediment currently used for the edge labels.  The main interest for
sediment is the number of times that the edge is taken.  A function,
extract_weight, has been added to extract that information from the
edge label.