Make sure heapq does not attempt to use the edge in tuples for ties breakers
heapq attempts to elminate ties by examining later elements in the
tuple.  Unfortunately, the type for edges does not have a compare
operator and trying to compare the causes an exception.  Putting in a
dummy counter that is incremented for every heapq.heappush to ensure
there are never any ties that would trigger the exception.